remoteproc: qcom_wcnss: Add qcom,wcn3660a compatible
WCN3660A is a variant of WCN3660, but with the same regulator
requirements as WCN3620/WCN3680/WCN3660B. As far as qcom_wcnss_iris
is concerned we can just use qcom,wcn3680 (wcn3680_data).

However, a separate compatible is needed for WCN3660A because
the wcn36xx driver uses it to enable chip-specific functionality.
In particular, it enables 802.11ac for qcom,wcn3680 which is not
supported by WCN3660A.

Signed-off-by: Matti Lehtimäki <matti.lehtimaki@gmail.com>