HACK!!! Do not attach clk to syscon
When the apcs syscon gets used, its first clock will be attached to the
syscon. When anything accesses the syscon then the clock will be
enabled. If this happens before the apcs-kpss-global driver is probed
then we'll have big issues reconfiguring the CPU PLL leading to crash of
device.

It seems the apcs-kpss-global driver should just probe much earlier to
avoid this, but for now let's just hack out this code so the a7pll clock
doesn't try and get reconfigured before everything's ready.